+++ src/components/header/header.jsx
@@ -21,7 +21,7 @@ import {
 } from './Header.styled';
 
 import AddFeedbackModal from './AddFeedbackModal/AddFeedbackModal';
-import { MobileSidebar } from './components/MobileSidebar';
+import { MobileSidebar } from './components/MobileSidebar/MobileSidebar';
 
 export const Header = () => {
   const location = useLocation();
